relocated all models to single models.py file and reran migrations
