+++ upoi/poi/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 
-from .models import Poi
+from .models import Poi, Tag
 
 admin.site.register(Poi)
+admin.site.register(Tag)
